Fix strange timing of a notification
Rewrote the code precisely according to the implementation of
'readProcess' solving the problem of 'waitforProcess'.
'waitforProcess' blocks all the other threads if a program is not
compiled with -threaded, but Xmonad cannot be compiled with -threaded,
see http://stackoverflow.com/a/14672588/2193620 .
The problem was that 'waitforProcess' blocked a thread spawn by
'withForkWait' and xmonad got deadlocked.